+++ 00-STM32_LIBRARIES/tm_stm32_usart.h
@@ -603,7 +603,7 @@ void TM_USART_Puts(USART_TypeDef* USARTx, char* str);
  * @param  count: Number of elements in data array to be send over USART
  * @retval None
  */
-void TM_USART_Send(USART_TypeDef* USARTx, uint8_t* DataArray, uint16_t count);
+void TM_USART_Send(USART_TypeDef* USARTx, uint8_t* DataArray, uint32_t count);
 
 /**
  * @brief  Gets character from internal USART buffer
